requirements.txt
requests
Flask
#gunicorn  # Si usas gunicorn
python-dotenv # Para cargar variables de entorno y desde .env